requirements.txt
google-genai
llama-index
llama-index-readers-file
python-dotenv
langgraph
langchain-google-genai
sentence-transformers
scikit-learn
ipython